Admin Permissions

+++ WebApp/views/guias/asistentesAdmin.php
@@ -8,19 +8,20 @@
 <main class="assistants_actions">
     <div class="asistentes">
         <h1 class="section__heading"><span>Asistentes Administrativas</span></h1>
-        <?php if(isAssistantCartago()){ ?>
+        <?php if(isAssistantCartago() || isAdmin()){ ?>
             <a href="guias/asignar/asistente" class="assistants_actions__type1">Asignar Asistente Administrativo por Campus </a>
             <br>
             <a href="/professors/register" class="assistants_actions__type2">Registrar Profesor</a>
             <br>
         <?php } ?>
-        <?php if(isAssistant()){ ?>
+        <?php if(isAssistant() || isAdmin()){ ?>
             <a href="guias/crear/equipo" class="assistants_actions__type1">Crear Equipo de Trabajo</a>
             <br>
             <a href="/ver/eliminar/equipo" class="assistants_actions__type2">Ver y Eliminar Equipo de Trabajo</a>
+            <br>
         <?php } ?>
 
-        <?php if(isTeacher()){ ?>
+        <?php if(isTeacher() && !isAssistant()){ ?>
             <a href="/ver/eliminar/equipo" class="assistants_actions__type2">Ver Equipo de Trabajo</a>
         <?php } ?>
 </div>
